# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{profileId}/meal-logs:
    get:
      operationId: listMealLogs
      tags: [meals]
      summary: List meal logs for a profile, optionally filtered by date
      parameters:
        - $ref: "#/components/parameters/ProfileIdParam"
        - name: date
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DateFilter"
      responses:
        "200":
          description: Meal logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealLog"
    post:
      operationId: createMealLog
      tags: [meals]
      summary: Log a food item eaten for a meal
      parameters:
        - $ref: "#/components/parameters/ProfileIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealLogInput"
      responses:
        "201":
          description: Created meal log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLog"
  /meal-logs/{mealLogId}:
    delete:
      operationId: deleteMealLog
      tags: [meals]
      summary: Remove a logged meal item
      parameters:
        - name: mealLogId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /profiles/{profileId}/lab-comparisons:
    get:
      operationId: listLabComparisons
      tags: [labs]
      summary: List baseline/follow-up lab comparisons for a profile
      parameters:
        - $ref: "#/components/parameters/ProfileIdParam"
      responses:
        "200":
          description: Lab comparisons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabComparison"
    post:
      operationId: createLabComparison
      tags: [labs]
      summary: Submit baseline and/or follow-up lab values and get progress insights
      parameters:
        - $ref: "#/components/parameters/ProfileIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabComparisonInput"
      responses:
        "201":
          description: Created lab comparison with insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabComparison"
components:
  parameters:
    ProfileIdParam:
      name: profileId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    DateFilter:
      type: string
      format: date
...
    MealLogInput:
      type: object
      required: [date, mealType, servings]
      properties:
        date:
          type: string
          format: date
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodId:
          type: integer
          nullable: true
          description: Dataset food id. Provide this OR customFoodName for a manual item.
        customFoodName:
          type: string
          description: Manual custom food name entered by the user (no dataset entry).
        customLabel:
          type: string
          description: Optional label/serving description for a manual custom food.
        servings:
          type: number
    MealLog:
      type: object
      required: [id, profileId, date, mealType, foodName, servings, protein, iron, calcium, vitaminD]
      properties:
        id:
          type: integer
        profileId:
          type: integer
        date:
          type: string
          format: date
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodId:
          type: integer
          nullable: true
        customFoodName:
          type: string
          nullable: true
        customLabel:
          type: string
          nullable: true
        foodName:
          type: string
        servings:
          type: number
        protein:
          type: number
        iron:
          type: number
        calcium:
          type: number
        vitaminD:
          type: number
...
    LabComparisonInput:
      type: object
      required: [labType]
      properties:
        labType:
          type: string
          enum: [baseline, followup]
        hemoglobin:
          type: ["number", "null"]
        ferritin:
          type: ["number", "null"]
        vitaminB12Level:
          type: ["number", "null"]
        vitaminDLevel:
          type: ["number", "null"]
        serumCalcium:
          type: ["number", "null"]
        totalProtein:
          type: ["number", "null"]
        rbcCount:
          type: ["number", "null"]
        wbcCount:
          type: ["number", "null"]
        plateletCount:
          type: ["number", "null"]
        hematocrit:
          type: ["number", "null"]
        mcv:
          type: ["number", "null"]
        serumIron:
          type: ["number", "null"]
        vitaminA:
          type: ["number", "null"]
        vitaminC:
          type: ["number", "null"]
        vitaminE:
          type: ["number", "null"]
        magnesium:
          type: ["number", "null"]
        phosphorus:
          type: ["number", "null"]
        sodium:
          type: ["number", "null"]
        fastingBloodSugar:
          type: ["number", "null"]
        hba1c:
          type: ["number", "null"]
        creatinine:
          type: ["number", "null"]
        bun:
          type: ["number", "null"]
        totalCholesterol:
          type: ["number", "null"]
        hdl:
          type: ["number", "null"]
        ldl:
          type: ["number", "null"]
        triglycerides:
          type: ["number", "null"]
        tsh:
          type: ["number", "null"]
        alt:
          type: ["number", "null"]
        ast:
          type: ["number", "null"]
        recordedAt:
          type: string
          format: date
    LabComparison:
      type: object
      required: [id, profileId, labType, recordedAt]
      properties:
        id:
          type: integer
        profileId:
          type: integer
        labType:
          type: string
          enum: [baseline, followup]
        hemoglobin:
          type: ["number", "null"]
        ferritin:
          type: ["number", "null"]
        vitaminB12Level:
          type: ["number", "null"]
        vitaminDLevel:
          type: ["number", "null"]
        serumCalcium:
          type: ["number", "null"]
        totalProtein:
          type: ["number", "null"]
        rbcCount:
          type: ["number", "null"]
        wbcCount:
          type: ["number", "null"]
        plateletCount:
          type: ["number", "null"]
        hematocrit:
          type: ["number", "null"]
        mcv:
          type: ["number", "null"]
        serumIron:
          type: ["number", "null"]
        vitaminA:
          type: ["number", "null"]
        vitaminC:
          type: ["number", "null"]
        vitaminE:
          type: ["number", "null"]
        magnesium:
          type: ["number", "null"]
        phosphorus:
          type: ["number", "null"]
        sodium:
          type: ["number", "null"]
        fastingBloodSugar:
          type: ["number", "null"]
        hba1c:
          type: ["number", "null"]
        creatinine:
          type: ["number", "null"]
        bun:
          type: ["number", "null"]
        totalCholesterol:
          type: ["number", "null"]
        hdl:
          type: ["number", "null"]
        ldl:
          type: ["number", "null"]
        triglycerides:
          type: ["number", "null"]
        tsh:
          type: ["number", "null"]
        alt:
          type: ["number", "null"]
        ast:
          type: ["number", "null"]
        recordedAt:
          type: string
          format: date
        insights:
          type: array
          items:
            type: string